+++ sdk/ai/azure-ai-projects/samples/evaluations/sample_evaluations.py
@@ -14,8 +14,8 @@ USAGE:
     Before running the sample:
 
     pip install azure-identity
-    pip install "git+https://github.com/Azure/azure-sdk-for-python.git@users/singankit/ai_project_utils#egg=azure-ai-client&subdirectory=sdk/ai/azure-ai-client"
-    pip install "git+https://github.com/Azure/azure-sdk-for-python.git@users/singankit/demo_evaluators_id#egg=azure-ai-evaluation&subdirectory=sdk/evaluation/azure-ai-evaluation"
+    pip install azure-ai-projects
+    pip install azure-ai-evaluation
 
     Set this environment variables with your own values:
     PROJECT_CONNECTION_STRING - the Azure AI Project connection string, as found in your AI Studio Project.