+++ sdk/monitor/azure-monitor-opentelemetry/azure/monitor/opentelemetry/_configure.py
@@ -7,6 +7,7 @@ from functools import cached_property
 from logging import getLogger
 from typing import Dict, List, cast
 
+from opentelemetry._events import set_event_logger_provider
 from opentelemetry._logs import set_logger_provider
 from opentelemetry.instrumentation.dependencies import (
     get_dist_dependency_conflicts,
@@ -15,6 +16,7 @@ from opentelemetry.instrumentation.instrumentor import (  # type: ignore
     BaseInstrumentor,
 )
 from opentelemetry.metrics import set_meter_provider
+from opentelemetry.sdk._events import EventLoggerProvider
 from opentelemetry.sdk._logs import LoggerProvider, LoggingHandler
 from opentelemetry.sdk._logs.export import BatchLogRecordProcessor
 from opentelemetry.sdk.metrics import MeterProvider
@@ -175,6 +177,10 @@ def _setup_logging(configurations: Dict[str, ConfigurationValue]):
     if not any(isinstance(handler, LoggingHandler) for handler in logger.handlers):
         handler = LoggingHandler(logger_provider=logger_provider)
         logger.addHandler(handler)
+
+    # Setup EventLoggerProvider
+    event_provider = EventLoggerProvider(logger_provider)
+    set_event_logger_provider(event_provider)
 
 
 def _setup_metrics(configurations: Dict[str, ConfigurationValue]):