+++ sdk/monitor/azure-monitor-opentelemetry/CHANGELOG.md
@@ -6,6 +6,8 @@
 
 - Only add OpenTelemetry LoggingHandler if current logger does not have it
     ([#38549](https://github.com/Azure/azure-sdk-for-python/pull/38549))
+- Distro to automatically configure event logger provider
+    ([#38543](https://github.com/Azure/azure-sdk-for-python/pull/38543))
 
 ### Breaking Changes